Adjust: Replace Colon with EM Dash Per Website
I just saw that blender.org/about has the exact same text verbatim, but their version correctly uses the em dash. Switching my PR here to em dash from colon so they are consistent. I also recently made a PR on the manual in this area and used EM dash there as well, with the exact same opening text verbatim. If both my PR's are approved, the "About Blender" section at the top of the About website, main README, and manual will all be perfectly consistent. Which is good because it is a very well-written opening blurb.