Fix NaN intersection results in edge cases
Since we used a distance check to determine one of the points to be inside the brush
and the other outside, it should be impossible to not have an intersection.
But certain cases still resulted in a negative discriminant, which
caused computations to fail.

Assuming all intersections to be at least tangential is a hacky solution, but
is likely the best we can do to salvage such cases without
aborting the cut completely.