Refactor: Replace "lockfree linklist" with Vector in icons code
Currently icon deletion is the only place using the lockfree linked
list implementation. It looks like it's used to avoid locking a mutex
when deleting an ID that has an associated icon. However, just
locking the mutex to append to a Vector should be just as fast if
not faster since it avoids an allocation that's currently necessary
for the linked list.

Replacing the linked list with a vector allows removing the
lockfree linked list utilities.