Fix #132042: Wrong IDname handling when moving a local ID into a library.
Mistake in 5f9f3116db2a, ID name needs to be removed from the runtime ID
name cache data _before_ it's library pointer is modified.

Library is an important part of an ID 'namesapce', so trying to remove
it after its library has been changed is trying to remove it from the
wrong namespace, leaving it in the original one.