Fix #127499: Adjust File/Asset Browser Preview Drag Area
Dragging assets in thumbnail view is made a bit more difficult because
we inset the hit area. This makes box select at the edges easier, but
at the expense of simple dragging.  We also inset in such a way that
you can't drag on the bottom-half of the name.  And the width of the
inset is related to the display size not resolution scale. This PR
changes this so that you can drag from almost the entirety of the tile
yet box select still works well enough.