Fix #131062: Active tool does not sync in Texture Painting workspace
Prior to this commit, after the brush asset project, bToolRef and
bToolRef_Runtime objects were not created or updated when the space type
between them differed. This is relevant for Texture Painting, where a
user may have both the 3D Viewport and the Image Editor open and the
active brush is expected to sync between them.

To support this functionality again, this commit makes the following
changes:

* Add an explicit method for texture paint to be called after activating
  the relevant brush asset
* Add helper methods which allocate and populate relevant data on the
  bToolRef_Runtime struct.