USD: update USDHook.py example
Updated the example to include new hook functions
material_import_poll and on_material_import as well
as the new `import_texture` and `export_texture`
utilities which were introduced in #131559.

Also reformatted the existing documentation to include
subheaders and internal links, as the previous paragraph
format became difficult to follow.