Fix #132264: Vertex Edge slide wrong
Caused by 580e0af309f2

The above commit fixed the condition that caused the hang in #128702 but
it also changed the `while` loop to a `do while`.
The do while loop executes the content of the loop once before checking
the condition of the while.
Whereas a while loop will check the condition first before executing the
content.

As a consequence, we could (falsely) enter the loop and return with a
(false) best direction.

To resolve, keep the check from 580e0af309f2 but go back to a `while`
loop (instead of the `do while).

Needs to go into 4.2 LTS