Do not bother checking if the vert is not an edge pair for taggging.  This extra processing provides no value.
If the vert is already an edge pair, and if dissolve will affect it, then it necessarily is a vert between a pair of faces.  The first face merge that dissolves one of the two edges will automatically remove both the other edge and also the vert too.  There's no reason to bother with counting edges and skipping tagging.

This was only required with the original logic, so that the operator would avoid dissolving entirely unreleated edge-pair verts that were already present in the mesh. (Which it did not actually do reliably.)

Since unselected verts are no longer candidates to be dissolved, the extra check is now uneeded.