WIP: Special Character Input for Windows Platform
On the Windows platform we process keyboard entry at a very low level
so we don't get any special character processing that is provided to
every other Windows program. Other platforms, like MacOS, are able
to allow pressing Alt-G to get a © for example, within Blender, but we
can't do that on Windows.  This PR allows special character input for
text inputs, Text Objects, Sequencer, and anywhere else we accept
keyboard entry.