Fix #132296: Transform: Use `transform` operator for dimentionless ops
For UV editor and sequencer preview, their translation value should be
dimentionless. This fix changes the keymap to use `transform.transform`
instead of `transform.translate` for those two places so the operator
will display a dimentionless translation value instead of showing with
units which will not align with user type-in value if the scene unit is
not set to Meters.