Fix #132325: Modelling: Use `DBL_EPSILON` when doing `is_quad_convex_v3`
While calculating the dot product of two arms of a quad in
`is_quad_convex_v3`, a small enough length is then discarded and the
function returns false (meaning quad is not convex). This would lead to
false negatives when quad is reasonably small but still convex in some
cases, leading to excessive triangulation when quads should be
preserved. Since `len_squared_v3` multiplies two `float`s, it makes
sense to use `DBL_EPSILON` instead of `FLT_EPSILON` to determine whether
the length is small enough.