Nodes: Add Menu: Add tooltip in "Group" submenu if no groups are available
This menu is empty if the current nodetree is not a group,
and there are no groups that can be added inside it.

Having an explicit tooltip that no groups are available
gives helpful context to the user and helps signal
that the menu being empty is not due to any glitches/errors.