Compositor: Add Set Realization Options node
This patch adds a new node to the compositor called the Set Realization
Options node. Its purpose is to tell nodes that follow it how they are
supposed to realize the image on their own domain. This now brings back
support for repetition and adjustable interpolation for realization.

The process of "Realization" happen when a node expects its inputs to
have a certain size, but the given inputs have a different size, so the
inputs need to be "realized" on the node's proffered size. See the code
documentation in COM_domain.hh for more information on how this happens.
This node allows the user to control how this realization happens.

The node provides two options:
- Interpolation: Which controls which interpolation method to use when
  realizing on the new size.
- Repeat: Which controls whether the image will repeat if the new size
  is larger than the image. If disabled, the empty areas will be filled
  with zeros.

Additionally, this node will hopefully make the concept of realization
clearer to the user, which used to be a rather confusing operation due
to its implicit nature. For instance, realization interpolation was only
possible to set using the Translate node, which is not really obvious.

For context, every since the introduction of the Full Frame compositor,
support for repeating images was removed and was rather difficult to
achieved using existing nodes, so user feedback made it clear that
bringing back support for repetition was paramount. But adding a node
specifically to repeat an image is also something to consider.