USD: Write extents out for Curves and more consistently for other types
We weren't writing extents out for Curves so do that now. Additionally,
make use of common code to write out the extents for meshes, hair,
lights, volumes, and points.

Move the curves test to live alongside its friends in Python rather than
C++ and add in the extents validation. Python tests allow for faster
iteration, are more straightforward to write, and allow usage of the USD
compliance validator for exports.