Spreadsheet: Move context path to footer, add editor menus
Currently the spreadsheet has no space for the typical editor
menus. Without them, we've had trouble adding new features
to the editor because there is no good place to expose them in
the UI.

This commit moves the context path to the left of the editor's
footer. In the leftover space editor menus are added, though we
only have a very short "View" menu right now.