Fix #81215: X11: Use OS Crosshair Cursor for Sculpt and Paint Modes
On X11 we use an OS-supplied mouse cursor (XC_crosshair) for Edit Mode
but for Sculpt, Vertex Paint, Weight Paint, and Texture Paint modes we
show WM_CURSOR_PAINT, which is GHOST_kStandardCursorCrosshairA. On
Windows and Mac this shows as a pretty nice cursor but on X11 this
shows one of our custom 16x16 aliased cursors which is a little clunky
and small. It is especially tiny on a 4K monitor. This PR uses the
nicer OS-supplied crosshair cursor for these other modes.