Fix #132562: Scrape Multiplane brush causes assert without samples
In some cases, using the Scrape Multiplane brush in Dynamic mode, the
brush stroke will have no vertices to sample from to find the associated
planes used to calculate deformations. This results in attempting to
divide by 0.

To make this potential error case more obvious, this commit changes the
helper method `sample_surface` to instead return a `std::optional` of
`ScrapeSampleData` in the rare occurance that we have no samples, so
that calling code is required to handle this case explicitly.