Cleanup: Make header files more clangd and clang-tidy friendly
This avoids seeing many errors when opening files with clangd as LSP
in an IDE, and makes it possible to do automatic clang-tidy fixes
from the command line.

* Add #includes and prototypes to make headers compile by themselves
* Add "IWYU pragma: export" in a few places
* Remove some unused #includes (but there are many more)
* Tweak ShaderCreateInfo macros to work with clangd

Some types of headers still have errors, these could be fixed or
worked around with more refactoring.

* Preprocesor template headers (e.g. NOD_static_types.h)
* Some inlining headers (e.g. CCGSubSurf_inline.h)

Note that that disablig WITH_UNITY_BUILD is required for clangd to
work properly, otherwise compile_commands.json can not be used to
find the commands for the united source files.