Refactor: RNA: add `discrete` suffix to `RNA_pointer_create`.
This is a noisy preliminary step to the 'RNA ancestors' change. The
rename helps clearly tell what each `pointer_create` function does.

---------

Split out into its own commit and PR as it is a very noisy change.

Only real potential point of discussion here is whether new proposed
name is good or not.