WIP: USD: Enable part of Import to be concurrent
Broadly speaking USD Import is separated into 2 main loops: Creating all
the Blender objects and then Loading all the data into those objects.
This change allows the second loop to be done in parallel as it does not
interact with the `bmain` database except in a few cases for materials
and armature.

Not using the `BKE_main_lock/unlock` APIs
These APIs use a spin lock, which is not well suited to the scenario
here. Spin locks are designed to protect a very small amount of code, on
the order of 20 instructions according to TBB, but typical bmain access
for writing often greatly exceeds that. For example, using these APIs to
protect the call to `BKE_image_load_exists` will lead to the import of
the ALab scene taking several minutes whereas a single-threaded import
takes ~7 seconds. The call to `ED_armature_from_edit` is similarly
affected.

Using a grain size of 1
This is set to 1 because the cost of each iteration is relatively high.
Trying with other values like  8, 16, and 24 showed little to no
difference in timing. However, values higher than that started to become
slower than the baseline here for me. A value of 1 allows other threads
to steal any item for processing if another thread is lagging behind.

Alternate design with temporary bmains
An alternate approach would be to use temporary, per-thread, bmains and
then merge them all at the very end. However, this wouldn't remove the
need for locking entirely and presents other problems. Locking is still
necessary to protect the various material maps we need to maintain. And
image/material de-duplication will still need to be "global" as these
are expensive to process and shouldn't wait until the merge at the end
to discover they're duplicates. More investigation might be required
for certain parenting and armature setups. It would be important that
each thread has access to all the datablocks it would need to establish
some parenting relationships and armature setups.

Results
Timing of the "second" loop (Before vs. After)
| ALab | Moore Lane 4004 | Caldera map_capital (full) |
| -- | -- | -- |
| 6.4 s vs. 4052ms | 3680ms vs 900ms | 70.2 s vs 5 s |

Note: ALab will drop closer to ~2 seconds once the Image TODOs below are
handled in some way.

----

TODO
To be done before checkin:
- usd_reader_pointinstancer: Determine if it's worth moving the GN node tree creation from `read_object_data` into `create_object` so we won't have to use the mutex there at all.

These can be done either before or after checkin:
- Image loading is almost 2x slower than it needs to be for ALab causing the lock spanning `BKE_image_load_exists` to be held for much longer than necessary. A new API will need to be created for more efficient UDIM image create+load scenarios.
- Future: `ED_armature_from_edit` is extremely inefficient due to how it continually searches through objects looking for those needing a call to BKE_pose_rebuild. This causes import to take 12+ minutes on one particular scene containing 805497 objects and 14218 armatures.