Fix #132734: Follow Active Quads fails on specific non-manifold mesh
So the thing is we have some manifold edges, the unwrap does not walk
further if it hits those.
However, the way the code checks if it needs to initialize edge lengths,
there is an assumption that edges has been visited already [which is
false in certain cases].

The easiest way to resolve is to check entries in `edge_lengths` for
**both** loop pairs, this way we are sure all entries are there when
needed.