Anim: handle library overrides when versioning legacy Actions
In this scenario:

- Working in Blender 4.3 or older,
- library file object already has animdata, but no Action assigned, and
- work file links the object, adds a library override to assign an
  Action.

opening the file in Blender 4.4 would:

- upgrade all the legacy Actions to slotted Actions, and
- ensure the slots get assigned wherever the legacy Action was assigned.

This would fail because the slot assignment wasn't stored in a library
override operation, and would thus get lost when the library overrides
are applied to the linked data.

This is now fixed by making the versioning code aware of library
overrides, and adding an override operation on the `action_slot_handle`
property.

Note: these library overrides will likely be removed by Blender again.
All slots created on just-versioned legacy Actions are named "Legacy
Slot", and the versioning code ensures that these are auto-assigned even
without override. The code for the auto-slot-assignment of NLA strips
and Action Constraints (see `blender::animrig::nla::assign_action` and
`rna_ActionConstraint_action_set`) is even more eager to auto-select a
slot whenever there is one that's suitable, and so it already picks up
on the Legacy Slot without the library overrides. And because this makes
the data the same as what is set in the library override, the
liboverride isn't necessary, and Blender removes it again.

To make this versioning code semantically correct without relying on the
specific implementation of code somewhere else in Blender, these
overrides are still created anyway.