Fix: invalid memory access in animrig::versioning unit tests
The issue was that we were passing a pointer-to-a-pointer to
`BLI_addtail()`, which expects a pointer to something castable to a
`Link`. This in turn led to an invalid memory access when trying to
access the fields of the supposed `Link`.

This fixes the issue by passing a pointer to a zero-initialized `Link`
instead.

Also took the opportunity to more simply zero-initialize the `bAction`
structs used in the tests as well.