Compositor: Add Glare Highlights Smoothness and Suppression
This patch adds two new inputs to the Glare node, Highlights Smoothness
and Highlights Suppression. Smoothness allows the user to control how
smooth the highlights are after thresholding and Suppression allows the
user to suppress very high brightness pixels.

Those are essentially similar to the Knee and Clamp options in old EEVEE
bloom, though they work differently.

The issue with the Knee parameter in old EEVEE bloom, aside from being
named after a body part, is that it actually isn't smooth or continuous
around zero if the threshold is sufficiently close to zero relative to
the Knee parameter. That's because zero lies in the smoothing kernel
region in those cases, and since zero pixels becoming highlights is very
bad, EEVEE just returned zero as a special case for zero brightness, but
values like 0.0001 will be full blown highlights.

The new nicely named Smoothness input uses adaptive smoothing such that
the smoothing kernel size will be reduced as the threshold nears zero,
such that smoothed highlights will be continuous and smooth around zero.

The Highlights Suppression input is similar to clamped, but it has a
[0, 1] range, and is relative to the brightest pixel in the image.

---

Since the node now has many inputs than before, it is probably better to
use node panels to separate the three functions of the node, Highlights
Extraction, Glare Generation, and Mixing. This will be done in a follow
up patch.